requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.16
django-environ==0.11.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-redis==5.4.0
Faker==30.8.2
loguru==0.7.2
psycopg2-binary==2.9.10
gunicorn==22.0.0
python-dateutil==2.9.0.post0
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
win32-setctime==1.1.0
